new and faster wordcount method
The old one took nearly a second for 120k words. The new one needs only 80-140ms. This is still not good in any way because it's still a O(n) algorithm that shouldn't be called if not needed (e.g. on every key type).
A better solution would be an additional progressive wordcount.