timed writing fixed & some formating fixes
Previously the time was uninitialized and had in most cases a really high value which lead
to a startTimer event with negative value. This was only visible as an error in the console
but it could also lead to the timer dialog popping up at random times.
Now there's only one timer and that timer will only be set if you click on "start" in the options
dialog. Drawback is that the timedWriting value will not be stored in the settings any longer.